# Utils: faster escaping functions (see issue #327)