See #327 
Bench results:

```bash
https://github.com/component/escape-html version:
  no special characters    x 22,456,447 ops/sec ±2.61% (192 runs sampled)
  single special character x  7,241,378 ops/sec ±1.32% (191 runs sampled)
  many special characters  x  3,542,471 ops/sec ±0.87% (192 runs sampled)

old bem-xjst 7.0.3 utils.xmlEscape version:
  no special characters    x  3,621,962 ops/sec ±2.34% (184 runs sampled)
  single special character x  2,796,355 ops/sec ±0.92% (192 runs sampled)
  many special characters  x  1,545,772 ops/sec ±0.86% (194 runs sampled)
```

I decided to update `utils.*Escape` functions.