@veged 2016-08-05
diff looks like copy-paste 3 times :-(