@miripiruni 2016-08-05
@veged interesting results I’ll just recived:

```
escape-html version:
  no special characters    x 24,746,854 ops/sec ±0.64% (195 runs sampled)
  single special character x  7,275,041 ops/sec ±2.54% (195 runs sampled)
  many special characters  x  3,778,574 ops/sec ±0.77% (193 runs sampled)
  many special characters  x     50,111 ops/sec ±0.60% (195 runs sampled)

current bem-xjst 7.x util.*Escape function:
  no special characters    x  3,905,172 ops/sec ±0.63% (194 runs sampled)
  single special character x  2,876,288 ops/sec ±0.78% (194 runs sampled)
  many special characters  x  1,580,096 ops/sec ±0.76% (194 runs sampled)
  many special characters  x    127,911 ops/sec ±1.55% (195 runs sampled)
```

The last one [is HTML from wikipedia article](https://github.com/miripiruni/escape-html/blob/master/benchmark/index.js#L47). As you see bem-xjst version is more  efficient.