@miripiruni 2016-08-05
> can you make benchmarks in terms of our «real» full bem-xjst usage?

Checking…