@veged 2016-08-05
it's will be cool to share many special characters result with author of https://github.com/component/escape-html