@miripiruni 2016-08-08
@veged  It’s doesn’t faster at all. In our test in bemjson there is nothing to escape. And `escapeContent` by default is `false`.

```
$ node run.js --compile --compare --min-samples 1000 --grep serp
compile:serp:next x 39.49 ops/sec ±1.95% (1018 runs sampled)
compile:serp:prev x 39.67 ops/sec ±2.43% (1021 runs sampled)
render:serp:next x 167 ops/sec ±1.68% (1027 runs sampled)
render:serp:prev x 168 ops/sec ±1.66% (1024 runs sampled)
```

More: 
```
$ node run.js --compile --compare --min-samples 1000
compile:basic:next x 994 ops/sec ±0.96% (1027 runs sampled)
compile:basic:prev x 975 ops/sec ±1.19% (1027 runs sampled)
render:basic:next x 342,235 ops/sec ±1.07% (1031 runs sampled)
render:basic:prev x 337,713 ops/sec ±1.07% (1032 runs sampled)

compile:granny:next x 185 ops/sec ±1.40% (1026 runs sampled)
compile:granny:prev x 190 ops/sec ±1.05% (1023 runs sampled)
render:granny:next x 399 ops/sec ±0.98% (1031 runs sampled)
render:granny:prev x 408 ops/sec ±0.94% (1029 runs sampled)

compile:islands:next x 351 ops/sec ±1.35% (1025 runs sampled)
compile:islands:prev x 360 ops/sec ±0.95% (1027 runs sampled)
render:islands:next x 704 ops/sec ±1.21% (1028 runs sampled)
render:islands:prev x 649 ops/sec ±0.75% (1027 runs sampled)

compile:serp:next x 42.02 ops/sec ±2.09% (1019 runs sampled)
compile:serp:prev x 40.39 ops/sec ±2.97% (1015 runs sampled)
render:serp:next x 148 ops/sec ±1.86% (1028 runs sampled)
render:serp:prev x 143 ops/sec ±1.93% (1024 runs sampled)

compile:showcase:next x 226 ops/sec ±3.08% (1020 runs sampled)
compile:showcase:prev x 250 ops/sec ±2.74% (1023 runs sampled)
render:showcase:next x 1,492 ops/sec ±1.33% (1023 runs sampled)
render:showcase:prev x 1,481 ops/sec ±1.10% (1034 runs sampled)
```