@veged 2016-08-08
10% is good enough! but we need to check with some sort of real life examples with `escapeContent:true` (which is gonna be default)