@miripiruni 2016-08-30
I created test with 1500 real `bemjson` from web4 prod.

Looks like even without `escapeContent: true` by default we can merge it, because of new escaping function faster in some cases and not slower in general: 

red — new escaping function
blue — old one

![screen shot 2016-08-30 at 15 33 43](https://cloud.githubusercontent.com/assets/144537/18089182/59300062-6ec7-11e6-8c5d-f9b42e488b3a.png)

Results with `escapeContent: true` 
![screen shot 2016-08-30 at 15 46 08](https://cloud.githubusercontent.com/assets/144537/18089494/f31c3172-6ec8-11e6-8c13-ac863dd8c324.png)