@miripiruni 2016-09-09
Without `escapeContent` option.

![faster-escaping-function-wo-escapecontent-option](https://cloud.githubusercontent.com/assets/144537/18370095/08e4c002-7633-11e6-942a-4e70698cec77.gif)



`3e80` — revision with new escaping function
`d536` — current HEAD.

Notice that all curve move to left. Because escaping function become faster and it affect all `bemhtml.apply(bemjson)` runs.

Case with `escapeContent:true` has not changed at all:

![faster-escaping-function-with-escapecontent-option](https://cloud.githubusercontent.com/assets/144537/18370955/6dd6a5ec-7639-11e6-8efd-ed7540e6ee60.gif)