@arikon 2016-09-09
@miripiruni В первой картинке хорошо бы максимум по оси Y уравнять. Сейчас там 1600 и 1800, и не очень наглядно видно разницу.