@miripiruni 2016-09-09
@arikon да, это gnuplot по умолчанию так нарезал ось. Масштаб прыгает как раз потому что некоторые результаты которы были в корзине ~40ms переползли в область 20—30ms.